pyproject.toml
[tool.poetry]
name = "jerome7"
version = "0.1.0"
description = "YU Show Up — 7 minutes a day. An act of love. A community of builders showing up."
authors = ["Omar Dominguez"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0"
alembic = "^1.14"
google-generativeai = "^0.8"
pydantic = "^2.0"
pydantic-settings = "^2.0"
click = "^8.1"
python-dotenv = "^1.0"
aiohttp = "^3.11"
"discord.py" = "^2.4"
jinja2 = "^3.1"
pillow = "^11.0"

[tool.poetry.scripts]
jerome7 = "src.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
